Change npm i to npm ci
Note that when you run `npm install` that will actually re-resolve all dependencies updating all of them to the most advanced versions supported. This means that what users will install will *not* be what is in the lockfile and therefore not what was tested by yall. The correct thing is `npm ci`